requirements.txt
aiolimiter==1.1.0
annotated-types==0.7.0
anyio==4.4.0
APScheduler==3.10.4
certifi==2024.7.4
cfgv==3.4.0
distlib==0.3.8
filelock==3.15.4
h11==0.14.0
httpcore==1.0.5
httpx==0.27.2
identify==2.6.0
idna==3.8
nodeenv==1.9.1
platformdirs==4.2.2
pre-commit==3.8.0
pydantic==2.8.2
pydantic-settings==2.4.0
pydantic_core==2.20.1
python-dotenv==1.0.1
python-telegram-bot==21.4
pytz==2024.1
PyYAML==6.0.2
six==1.16.0
sniffio==1.3.1
typing_extensions==4.12.2
tzlocal==5.2
virtualenv==20.26.3
